style: refactor to be more readable
