+++ lightning_api/README.md
@@ -1,4 +1,4 @@
 # LightningAPI
 
 LightningAPI is a thin wrapper around lightning APIs that are used for the agents (currently it just supports the minimal that is needed for [channelchecker](../channelchecker))
-We have LND gRPC and REST support for the time being.
+We have LND gRPC, LND REST and CLN (over local UNIX domain socket) support for the time being.
