+++ lightning_api/api_clnsocket.go
@@ -3,273 +3,46 @@ package lightning_api
 import (
 	"context"
 	"fmt"
-	"math/big"
-	"strconv"
-	"strings"
+	"os"
 	"time"
+
+	r "net/rpc"
 
 	entities "github.com/bolt-observer/go_common/entities"
 	"github.com/golang/glog"
 	rpc "github.com/powerman/rpc-codec/jsonrpc2"
 )
 
-var (
-	ErrNoNode          = fmt.Errorf("node not found")
-	ErrNoChan          = fmt.Errorf("channel not found")
-	ErrInvalidResponse = fmt.Errorf("invalid response")
+// Compile time check for the interface
+var _ LightingApiCalls = &ClnSocketLightningApi{}
+
+const (
+	LISTCHANNELS = "listchannels"
+	LISTNODES    = "listnodes"
+	LISTFUNDS    = "listfunds"
+	GETINFO      = "getinfo"
 )
 
-type ClnInfo struct {
-	PubKey    string            `json:"id"`
-	Alias     string            `json:"alias"`
-	Color     string            `json:"color"`
-	Network   string            `json:"network"`
-	Addresses []ClnListNodeAddr `json:"address,omitempty"`
-	Features  ClnFeatures       `json:"our_features"`
-}
-
-type ClnFeatures struct {
-	Init    string `json:"init"`
-	Node    string `json:"node"`
-	Channel string `json:"channe"`
-	Invoice string `json:"invoice"`
-}
-
-type ClnSetChan struct {
-	PeerId         string `json:"peer_id"`
-	LongChanId     string `json:"channel_id"`
-	FeeBase        string `json:"fee_base_msat"`
-	FeePpm         string `json:"fee_proportional_milli"`
-	MiHtlc         string `json:"minimum_htlc_out_msat"`
-	MaxHtlc        string `json:"maximum_htlc_out_msat"`
-	ShortChannelId string `json:"short_channel_id,omitempty"`
-}
-
-type ClnSetChanResp struct {
-	Settings []ClnSetChan `json:"channels,omitempty"`
-}
-
-type ClnListChan struct {
-	Source         string            `json:"source"`
-	Destination    string            `json:"destination"`
-	Public         bool              `json:"public"`
-	Capacity       uint64            `json:"satoshis"`
-	Active         bool              `json:"active"`
-	LastUpdate     entities.JsonTime `json:"last_update"`
-	FeeBase        uint64            `json:"base_fee_millisatoshi"`
-	FeePpm         uint64            `json:"fee_per_millionth"`
-	MinHtlc        string            `json:"htlc_minimum_msat"`
-	MaxHtlc        string            `json:"htlc_maximum_msat"`
-	ShortChannelId string            `json:"short_channel_id,omitempty"`
-	Delay          uint64            `json:"delay"`
-}
-
-type ClnListChanResp struct {
-	Channels []ClnListChan `json:"channels,omitempty"`
-}
-
-type ClnFundsChan struct {
-	PeerId         string `json:"peer_id"`
-	Connected      bool   `json:"connected,omitempty"`
-	ShortChannelId string `json:"short_channel_id"`
-	State          string `json:"state"`
-	Capacity       uint64 `json:"channel_total_sat"`
-	OurAmount      uint64 `json:"channel_sat"`
-	FundingTxId    string `json:"funding_tx_id"`
-	FundingOutput  int    `json:"funding_output"`
-}
-
-type ClnFundsChanResp struct {
-	Channels []ClnFundsChan `json:"channels,omitempty"`
-}
-
-type ClnSocketLightningApi struct {
-	LightningApi
-	Client *rpc.Client
-}
-
-type ClnListNodeAddr struct {
-	Type    string `json:"type"`
-	Address string `json:"address"`
-	Port    int    `json:"port"`
-}
-
-type ClnListNode struct {
-	PubKey     string             `json:"nodeid"`
-	Alias      string             `json:"alias,omitempty"`
-	Color      string             `json:"color,omitempty"`
-	Features   string             `json:"features,omitempty"`
-	Addresses  []ClnListNodeAddr  `json:"addresses,omitempty"`
-	LastUpdate *entities.JsonTime `json:"last_update,omitempty"`
-}
-
-type ClnListNodeResp struct {
-	Nodes []ClnListNode `json:"nodes,omitempty"`
-}
-
-func ToLndChanId(id string) (uint64, error) {
-
-	split := strings.Split(strings.ToLower(id), "x")
-	if len(split) != 3 {
-		return 0, fmt.Errorf("wrong channel id: %v", id)
-	}
-
-	blockId, err := strconv.ParseUint(split[0], 10, 64)
+// Usage "unix", "/home/ubuntu/.lightning/bitcoin/lightning-rpc"
+func NewClnSocketLightningApiRaw(socketType string, address string) LightingApiCalls {
+	client, err := rpc.Dial(socketType, address)
 	if err != nil {
-		return 0, err
+		glog.Warningf("Got error: %v", err)
+		return nil
 	}
-
-	txIdx, err := strconv.ParseUint(split[1], 10, 64)
-	if err != nil {
-		return 0, err
-	}
-
-	outputIdx, err := strconv.ParseUint(split[2], 10, 64)
-	if err != nil {
-		return 0, err
-	}
-
-	result := (blockId&0xffffff)<<40 | (txIdx&0xffffff)<<16 | (outputIdx & 0xffff)
-
-	return result, nil
+	return &ClnSocketLightningApi{Client: client, Timeout: time.Second * 30}
 }
 
-func FromLndChanId(chanId uint64) string {
-	blockId := int64((chanId & 0xffffff0000000000) >> 40)
-	txIdx := int((chanId & 0x000000ffffff0000) >> 16)
-	outputIdx := int(chanId & 0x000000000000ffff)
-
-	return fmt.Sprintf("%dx%dx%d", blockId, txIdx, outputIdx)
-}
-
-func ConvertAmount(s string) uint64 {
-	x := strings.ReplaceAll(s, "msat", "")
-	ret, err := strconv.ParseUint(x, 10, 64)
-	if err != nil {
-		glog.Warningf("Could not convert: %v %v", s, err)
-		return 0
+func NewClnSocketLightningApi(getData GetDataCall) LightingApiCalls {
+	if getData == nil {
+		return nil
 	}
-
-	return ret
-}
-
-func ConvertFeatures(features string) map[string]NodeFeatureApi {
-	n := new(big.Int)
-
-	n, ok := n.SetString(features, 16)
-	if !ok {
+	data, err := getData()
+	if data == nil || err != nil {
 		return nil
 	}
 
-	result := make(map[string]NodeFeatureApi)
-
-	m := big.NewInt(0)
-	zero := big.NewInt(0)
-	two := big.NewInt(2)
-
-	bit := 0
-	for n.Cmp(zero) == 1 {
-		n.DivMod(n, two, m)
-
-		if m.Cmp(zero) != 1 {
-			// Bit is not set
-			bit++
-			continue
-		}
-
-		result[fmt.Sprintf("%d", bit)] = NodeFeatureApi{
-			Name:       "",
-			IsKnown:    true,
-			IsRequired: bit%2 == 0,
-		}
-
-		bit++
-	}
-
-	return result
-}
-
-func (l *ClnSocketLightningApi) GetInternalChannels(pubKey string) (map[string][]ClnListChan, error) {
-	result := make(map[string][]ClnListChan, 0)
-
-	var listChanReply ClnListChanResp
-	err := l.Client.Call("listchannels", []interface{}{nil, pubKey, nil}, &listChanReply)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, one := range listChanReply.Channels {
-		if one.ShortChannelId == "" {
-			continue
-		}
-
-		if _, ok := result[one.ShortChannelId]; !ok {
-			result[one.ShortChannelId] = make([]ClnListChan, 0)
-		}
-
-		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
-	}
-
-	err = l.Client.Call("listchannels", []interface{}{nil, nil, pubKey}, &listChanReply)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, one := range listChanReply.Channels {
-		if one.ShortChannelId == "" {
-			continue
-		}
-
-		if _, ok := result[one.ShortChannelId]; !ok {
-			result[one.ShortChannelId] = make([]ClnListChan, 0)
-		}
-
-		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
-	}
-
-	return result, nil
-}
-
-func (l *ClnSocketLightningApi) GetInternalChannelsAll() (map[string][]ClnListChan, error) {
-	result := make(map[string][]ClnListChan, 0)
-
-	var listChanReply ClnListChanResp
-	err := l.Client.Call("listchannels", []interface{}{}, &listChanReply)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, one := range listChanReply.Channels {
-		if one.ShortChannelId == "" {
-			continue
-		}
-
-		if _, ok := result[one.ShortChannelId]; !ok {
-			result[one.ShortChannelId] = make([]ClnListChan, 0)
-		}
-
-		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
-	}
-
-	return result, nil
-}
-
-func SumCapacitySimple(channels []NodeChannelApi) uint64 {
-	sum := uint64(0)
-	for _, channel := range channels {
-		sum += channel.Capacity
-	}
-
-	return sum
-}
-
-func SumCapacityExtended(channels []NodeChannelApiExtended) uint64 {
-	sum := uint64(0)
-	for _, channel := range channels {
-		sum += channel.Capacity
-	}
-
-	return sum
+	return NewClnSocketLightningApiRaw("unix", data.Endpoint)
 }
 
 func (l *ClnSocketLightningApi) Cleanup() {
@@ -278,7 +51,7 @@ func (l *ClnSocketLightningApi) Cleanup() {
 
 func (l *ClnSocketLightningApi) DescribeGraph(ctx context.Context, unannounced bool) (*DescribeGraphApi, error) {
 	var reply ClnListNodeResp
-	err := l.Client.Call("listnodes", []interface{}{}, &reply)
+	err := l.CallWithTimeout(LISTNODES, []interface{}{}, &reply)
 	if err != nil {
 		return nil, err
 	}
@@ -383,7 +156,7 @@ func ConvertChannelInternal(chans []ClnListChan, id uint64) (*NodeChannelApiExte
 func (l *ClnSocketLightningApi) GetChanInfo(ctx context.Context, chanId uint64) (*NodeChannelApi, error) {
 
 	var listChanReply ClnListChanResp
-	err := l.Client.Call("listchannels", []string{FromLndChanId(chanId)}, &listChanReply)
+	err := l.CallWithTimeout(LISTCHANNELS, []string{FromLndChanId(chanId)}, &listChanReply)
 	if err != nil {
 		return nil, err
 	}
@@ -403,7 +176,7 @@ func (l *ClnSocketLightningApi) GetChanInfo(ctx context.Context, chanId uint64) 
 
 func (l *ClnSocketLightningApi) GetChannels(ctx context.Context) (*ChannelsApi, error) {
 	var fundsReply ClnFundsChanResp
-	err := l.Client.Call("listfunds", []string{}, &fundsReply)
+	err := l.CallWithTimeout(LISTFUNDS, []string{}, &fundsReply)
 	if err != nil {
 		return nil, err
 	}
@@ -414,7 +187,7 @@ func (l *ClnSocketLightningApi) GetChannels(ctx context.Context) (*ChannelsApi, 
 
 	for _, one := range fundsReply.Channels {
 
-		err = l.Client.Call("listchannels", []string{one.ShortChannelId}, &listChanReply)
+		err = l.CallWithTimeout(LISTCHANNELS, []string{one.ShortChannelId}, &listChanReply)
 		if err != nil {
 			return nil, err
 		}
@@ -461,7 +234,7 @@ func (l *ClnSocketLightningApi) GetChannels(ctx context.Context) (*ChannelsApi, 
 
 func (l *ClnSocketLightningApi) GetInfo(ctx context.Context) (*InfoApi, error) {
 	var reply ClnInfo
-	err := l.Client.Call("getinfo", []string{}, &reply)
+	err := l.CallWithTimeout(GETINFO, []string{}, &reply)
 	if err != nil {
 		return nil, err
 	}
@@ -475,7 +248,6 @@ func (l *ClnSocketLightningApi) GetInfo(ctx context.Context) (*InfoApi, error) {
 }
 
 func ConvertNodeInfo(node ClnListNode) *DescribeGraphNodeApi {
-
 	if node.LastUpdate == nil {
 		glog.Warningf("Gossip message for node %s was not received yet", node.PubKey)
 		return &DescribeGraphNodeApi{
@@ -517,7 +289,7 @@ func ConvertAddresses(addr []ClnListNodeAddr) []NodeAddressApi {
 
 func (l *ClnSocketLightningApi) GetNodeInfo(ctx context.Context, pubKey string, channels bool) (*NodeInfoApi, error) {
 	var reply ClnListNodeResp
-	err := l.Client.Call("listnodes", []string{pubKey}, &reply)
+	err := l.CallWithTimeout(LISTNODES, []string{pubKey}, &reply)
 	if err != nil {
 		return nil, err
 	}
@@ -572,7 +344,7 @@ func (l *ClnSocketLightningApi) GetNodeInfo(ctx context.Context, pubKey string, 
 
 func (l *ClnSocketLightningApi) GetNodeInfoFull(ctx context.Context, channels bool, unannounced bool) (*NodeInfoApiExtended, error) {
 	var reply ClnInfo
-	err := l.Client.Call("getinfo", []string{}, &reply)
+	err := l.CallWithTimeout(GETINFO, []string{}, &reply)
 	if err != nil {
 		return nil, err
 	}
@@ -633,27 +405,78 @@ func (l *ClnSocketLightningApi) GetNodeInfoFull(ctx context.Context, channels bo
 	return result, nil
 }
 
-// Compile time check for the interface
-var _ LightingApiCalls = &ClnSocketLightningApi{}
+func (l *ClnSocketLightningApi) GetInternalChannels(pubKey string) (map[string][]ClnListChan, error) {
+	result := make(map[string][]ClnListChan, 0)
 
-// Usage "unix", "/home/ubuntu/.lightning/bitcoin/lightning-rpc"
-func NewClnSocketLightningApiRaw(socketType string, address string) LightingApiCalls {
-	client, err := rpc.Dial(socketType, address)
+	var listChanReply ClnListChanResp
+	err := l.CallWithTimeout(LISTCHANNELS, []interface{}{nil, pubKey, nil}, &listChanReply)
 	if err != nil {
-		glog.Warningf("Got error: %v", err)
-		return nil
+		return nil, err
 	}
-	return &ClnSocketLightningApi{Client: client}
+
+	for _, one := range listChanReply.Channels {
+		if one.ShortChannelId == "" {
+			continue
+		}
+
+		if _, ok := result[one.ShortChannelId]; !ok {
+			result[one.ShortChannelId] = make([]ClnListChan, 0)
+		}
+
+		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
+	}
+
+	err = l.CallWithTimeout(LISTCHANNELS, []interface{}{nil, nil, pubKey}, &listChanReply)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, one := range listChanReply.Channels {
+		if one.ShortChannelId == "" {
+			continue
+		}
+
+		if _, ok := result[one.ShortChannelId]; !ok {
+			result[one.ShortChannelId] = make([]ClnListChan, 0)
+		}
+
+		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
+	}
+
+	return result, nil
 }
 
-func NewClnSocketLightningApi(getData GetDataCall) LightingApiCalls {
-	if getData == nil {
-		return nil
-	}
-	data, err := getData()
-	if data == nil || err != nil {
-		return nil
+func (l *ClnSocketLightningApi) GetInternalChannelsAll() (map[string][]ClnListChan, error) {
+	result := make(map[string][]ClnListChan, 0)
+
+	var listChanReply ClnListChanResp
+	err := l.CallWithTimeout(LISTCHANNELS, []interface{}{}, &listChanReply)
+	if err != nil {
+		return nil, err
 	}
 
-	return NewClnSocketLightningApiRaw("unix", data.Endpoint)
+	for _, one := range listChanReply.Channels {
+		if one.ShortChannelId == "" {
+			continue
+		}
+
+		if _, ok := result[one.ShortChannelId]; !ok {
+			result[one.ShortChannelId] = make([]ClnListChan, 0)
+		}
+
+		result[one.ShortChannelId] = append(result[one.ShortChannelId], one)
+	}
+
+	return result, nil
+}
+
+func (l *ClnSocketLightningApi) CallWithTimeout(serviceMethod string, args any, reply any) error {
+	c := make(chan *r.Call, 1)
+	go func() { l.Client.Go(serviceMethod, args, reply, c) }()
+	select {
+	case call := <-c:
+		return call.Error
+	case <-time.After(l.Timeout):
+		return os.ErrDeadlineExceeded
+	}
 }
